added circular singly linked list with basic methods
`Node` class represents an individual node within the linked list.
`List` class provides methods for creating a new list.

Added methods:
`length()`: returns the size of the list.
`append(value)`: adds an element to the end of the list.
`insert(value, index)`: inserts a new element at a given index in the list.
`get(index)`: returns the value of the element at a given index in the list.