changed implementation to be based on build-in arrays
Old implementation (circular singly linked list) is a dynamic data structure that can easily grow or shrink as items are added or removed. However, it requires more overhead in terms of memory and time complexity, as each node must store a reference to the next node.

New (array-based) implementation is simpler and more efficient in terms of memory usage and time complexity. However, it may not be as flexible as the linked list approach, as the size of the array is fixed and may need to be resized if the number of items in the list changes.